Config classes updated

+++ shop-miner/src/main/java/br/ufrn/shopminer/service/ConfigService.java
@@ -31,7 +31,11 @@ public class ConfigService {
 	}
 
 	@Transactional(readOnly = false)
-	public void delete(Config entity) {
-		configRepository.delete(entity);
+	public void delete(Config entity) throws Exception {
+		if (entity.getSites() != null)
+			configRepository.delete(entity);
+		else {
+			throw new Exception("Cannot delete. Config is not empty");
+		}
 	}
 }
